+++ src/port.h
@@ -49,7 +49,7 @@ int8_t write_saved_soc_group(float grpsoc);
  * @return  0: valid
  *         -1: invalid
  */
-int8_t read_saved_soh(double *soh);
+int8_t read_saved_soh(float *soh);
 
 /** 
  * @brief write lastest soh (per 1% changed)
@@ -57,7 +57,7 @@ int8_t read_saved_soh(double *soh);
  * @return  0: valid
  *         -1: invalid
  */
-int8_t write_saved_soh(double *soh);
+int8_t write_saved_soh(float *soh);
 
 /** 
  * @brief read saved cycle (last cycle before shutdown)
