+++ src/soc.c
@@ -290,11 +290,11 @@ void mysoc_pureAH(struct SOC_Info *SOCinfo, float cur, uint16_t vol, uint16_t te
 {
     const uint16_t cap = get_cap(cur, tempra);
 
-    const double capf = cap/10.0*(soh/100);
+    const float capf = cap/10.0*(soh/100);
 
-    double diffAH = DIFF_T_SEC/3600.0*cur/capf*100;
-    double SOCcal = SOCinfo->soc + diffAH;
-    double SOCer2Cal = SOCinfo->socEr2 + EKF_Q(diffAH,capf, cur);
+    float diffAH = DIFF_T_SEC/3600.0*cur/capf*100;
+    float SOCcal = SOCinfo->soc + diffAH;
+    float SOCer2Cal = SOCinfo->socEr2 + EKF_Q(diffAH,capf, cur);
 
 
 
@@ -356,7 +356,7 @@ void mysoc_smooth(struct SOC_Info *SOCinfo, float cur, uint16_t vol, uint16_t te
             
         }else if(cur<0 && ((int)oldest_vol-(int)newest_vol)){
             int smooth_empty_estimate_s = (vol-*g_dsg_stop_vol)/(oldest_vol-newest_vol)*(CELL_VOL_BUFFER_LEN-1)*CELL_VOL_BUFFER_SAMPLE_TIME_S;
-            double smoothDiff = SOCinfo->soc_smooth/smooth_empty_estimate_s*(3+15*(vol-*g_dsg_stop_vol)/(SOC_SMOOTH_START_VOL_DSG-*g_dsg_stop_vol));
+            float smoothDiff = SOCinfo->soc_smooth/smooth_empty_estimate_s*(3+15*(vol-*g_dsg_stop_vol)/(SOC_SMOOTH_START_VOL_DSG-*g_dsg_stop_vol));
             SOCinfo->soc_smooth -= smoothDiff;
             if(callcount%16 == 8 && SOCinfo->soc_smooth<1){
                 printf("dsg soc speedup %f,smooth_full_estimate_s: %d, soc_smooth %f\r\n", smoothDiff, smooth_empty_estimate_s, SOCinfo->soc_smooth);
@@ -375,7 +375,7 @@ void mysoc_smooth(struct SOC_Info *SOCinfo, float cur, uint16_t vol, uint16_t te
 }
 
 
-double getEKF_Q(double soc)
+float getEKF_Q(float soc)
 {
     if(g_group_state == GROUP_STATE_charging)
     {
@@ -389,7 +389,7 @@ double getEKF_Q(double soc)
 
 
 
-uint32_t getEKF_R(double H)
+uint32_t getEKF_R(float H)
 {
 
     if(H<1)
@@ -413,51 +413,51 @@ void mysocEKF(struct SOC_Info *SOCinfo, float cur, uint16_t vol, uint16_t tempra
 {
     static int callCount = 0;
     callCount++;
-    static double pureAHSUM = 0;
+    static float pureAHSUM = 0;
 
     const uint16_t *curve = get_curve_v(cur, tempra);
     const int16_t *curveK = get_curve_k(cur, tempra);
     const uint16_t cap = get_cap(cur, tempra);
 
-    const double capf = cap/10.0*(soh/100);
+    const float capf = cap/10.0*(soh/100);
 
-    double diffAH = DIFF_T_SEC/3600.0*cur/capf*100;
-    double SOCcal = SOCinfo->soc + diffAH;
-    double Q = getEKF_Q(SOCcal);
-    double SOCer2Cal = SOCinfo->socEr2 + Q;
+    float diffAH = DIFF_T_SEC/3600.0*cur/capf*100;
+    float SOCcal = SOCinfo->soc + diffAH;
+    float Q = getEKF_Q(SOCcal);
+    float SOCer2Cal = SOCinfo->socEr2 + Q;
     pureAHSUM += diffAH;
     // printf("diffAH : %f, EKF_W : %f pureAH: %f\n", diffAH, EKF_W(diffAH,capf, cur), pureAHSUM);
-    double H = 0, Hprev = 0, Hnext = 0;
-    double estVol = 0, estVolPrev = 0, estVolNext = 0;
-    double SOCerCal = sqrt(SOCer2Cal);
+    float H = 0, Hprev = 0, Hnext = 0;
+    float estVol = 0, estVolPrev = 0, estVolNext = 0;
+    float SOCerCal = sqrt(SOCer2Cal);
     if(SOCcal < 0)
     {
         SOCcal = 0;
 
         estVol = curve[0];
-        H = (double)curveK[0]/10;
+        H = (float)curveK[0]/10;
     }else if(SOCcal > 100)
     {
         SOCcal = 100;
 
         estVol = curve[SOC_POINT_NUM-1];
-        H = (double)curveK[SOC_POINT_NUM-1]/10;
+        H = (float)curveK[SOC_POINT_NUM-1]/10;
     }else{
         if(SOCcal < 5)
         {
-            double sock = SOCcal-(int)SOCcal;
+            float sock = SOCcal-(int)SOCcal;
             estVolPrev = curve[(int)SOCcal];
             estVolNext = curve[(int)SOCcal+1];
             estVol = estVolPrev + sock*(estVolNext-estVolPrev);
         }else if(SOCcal < 95)
         {
-            double sock = ((int)SOCcal%5+SOCcal-(int)SOCcal)/5;
+            float sock = ((int)SOCcal%5+SOCcal-(int)SOCcal)/5;
             estVolPrev = curve[(int)SOCcal/5+4];
             estVolNext = curve[(int)SOCcal/5+1+4];
             estVol = estVolPrev + sock*(estVolNext-estVolPrev);
         }    
         else{
-            double sock = SOCcal-(int)SOCcal;
+            float sock = SOCcal-(int)SOCcal;
             estVolPrev = curve[23+(int)SOCcal-95];
             estVolNext = curve[23+(int)SOCcal-95+1];
             estVol = estVolPrev + sock*(estVolNext-estVolPrev);
@@ -466,26 +466,26 @@ void mysocEKF(struct SOC_Info *SOCinfo, float cur, uint16_t vol, uint16_t tempra
 
         if(SOCcal < 5)
         {
-            double Hk = SOCcal-(int)SOCcal;
-            Hprev = (double)curveK[((int)SOCcal)]/10;
-            Hnext = (double)curveK[(int)SOCcal+1]/10;
+            float Hk = SOCcal-(int)SOCcal;
+            Hprev = (float)curveK[((int)SOCcal)]/10;
+            Hnext = (float)curveK[(int)SOCcal+1]/10;
             H = Hprev + Hk*(Hnext-Hprev); 
         }else if(SOCcal < 95)
         {
-            double Hk = ((int)SOCcal%5+SOCcal-(int)SOCcal)/5;
-            Hprev = (double)curveK[((int)SOCcal/5+4)]/10;
-            Hnext = (double)curveK[(int)SOCcal/5+1+4]/10;
+            float Hk = ((int)SOCcal%5+SOCcal-(int)SOCcal)/5;
+            Hprev = (float)curveK[((int)SOCcal/5+4)]/10;
+            Hnext = (float)curveK[(int)SOCcal/5+1+4]/10;
             H = Hprev + Hk*(Hnext-Hprev); 
         }    
         else{
-            double Hk = SOCcal-(int)SOCcal;
-            Hprev = (double)curveK[23+(int)SOCcal-95]/10;
-            Hnext = (double)curveK[23+(int)SOCcal-95+1]/10;
+            float Hk = SOCcal-(int)SOCcal;
+            Hprev = (float)curveK[23+(int)SOCcal-95]/10;
+            Hnext = (float)curveK[23+(int)SOCcal-95+1]/10;
             H = Hprev + Hk*(Hnext-Hprev); 
         }
         // printf("callcount %d hprev :%f  H : %f  hnext :%f \n", callCount, Hprev, H, Hnext);
     }
-    double K = 0;
+    float K = 0;
     uint32_t ekfR = getEKF_R(H);
 
     // if(callCount/16 == 2000){
@@ -498,14 +498,14 @@ void mysocEKF(struct SOC_Info *SOCinfo, float cur, uint16_t vol, uint16_t tempra
     K = SOCer2Cal*H/(H*SOCer2Cal*H+ekfR);
 
     
-    double res = SOCcal+K*((double)vol-estVol);
-    // printf("callcount %d  H: %f K :%f  kcal : %f , vol: %d, estvol: %lf\n", callCount, H, K, K*((double)vol-estVol), vol, estVol);
-    double resEr2 = (1-K*H)*SOCer2Cal;
+    float res = SOCcal+K*((float)vol-estVol);
+    // printf("callcount %d  H: %f K :%f  kcal : %f , vol: %d, estvol: %lf\n", callCount, H, K, K*((float)vol-estVol), vol, estVol);
+    float resEr2 = (1-K*H)*SOCer2Cal;
     if(resEr2 > Q)
     {
         resEr2 = Q;
     }
-    double SOCerRes = sqrt(resEr2);
+    float SOCerRes = sqrt(resEr2);
 
 
     if(res < 0)
@@ -524,7 +524,7 @@ void mysocEKF(struct SOC_Info *SOCinfo, float cur, uint16_t vol, uint16_t tempra
             //res = SOCinfo->soc;
             // resEr2 = SOCinfo->socEr2;
             // printf("soc underflow\n");
-            double smooth_k = 10/SOCinfo->soc;
+            float smooth_k = 10/SOCinfo->soc;
             if(smooth_k > 1){
                 smooth_k = 1;
             }
@@ -535,7 +535,7 @@ void mysocEKF(struct SOC_Info *SOCinfo, float cur, uint16_t vol, uint16_t tempra
         if(res > SOCinfo->soc)
         {
             //res = SOCinfo->soc;
-            double smooth_k = 10/(100-SOCinfo->soc);
+            float smooth_k = 10/(100-SOCinfo->soc);
             if(smooth_k > 1){
                 smooth_k = 1;
             }
@@ -555,7 +555,7 @@ void mysocEKF(struct SOC_Info *SOCinfo, float cur, uint16_t vol, uint16_t tempra
     SOCinfo->socEr2 = resEr2;
     // printf("soc : %f \n", 100-res);
     if(callCount%16 == 0){
-        printf("%d soc error2: %f Q:%f R:%d, H:%f, K:%lf kcal : %f , vol: %d, estvol: %lf\n",callCount/16, SOCinfo->socEr2, Q, ekfR, H, K, K*((double)vol-estVol), vol, estVol);
+        printf("%d soc error2: %f Q:%f R:%d, H:%f, K:%lf kcal : %f , vol: %d, estvol: %lf\n",callCount/16, SOCinfo->socEr2, Q, ekfR, H, K, K*((float)vol-estVol), vol, estVol);
     }
     
 }
@@ -628,16 +628,16 @@ static void bubbleSort_ascend_float(float *inputArr, float *outputArr, uint16_t 
     }
 }
 
-static void bubbleSort_ascend_duble(double *inputArr, double *outputArr, uint16_t size)
+static void bubbleSort_ascend_duble(float *inputArr, float *outputArr, uint16_t size)
 {
-    memcpy(outputArr, inputArr, size*sizeof(double));
+    memcpy(outputArr, inputArr, size*sizeof(float));
     for (size_t i = 0; i < size-1; i++)
     {
         for (size_t j = 0; j < size-i-1; j++)
         {
             if(outputArr[j] > outputArr[j+1])
             {
-                double tmp = outputArr[j];
+                float tmp = outputArr[j];
                 outputArr[j] = outputArr[j+1];
                 outputArr[j+1] = tmp;
             }
